productivitytracker-server/requirements.txt
appnope==0.1.0
attrs==17.3.0
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
decorator==4.1.2
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
Faker==0.7.3
gunicorn==19.7.1
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
mixer==5.6.6
oauthlib==2.0.6
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
Pygments==2.2.0
PyJWT==1.5.3
pytest==3.3.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-ipdb==0.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.0
